requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
apollo==4.2.13
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.1.0
bcbio-gff==0.7.0
biopython==1.81
CacheControl==0.12.11
cachetools==3.1.1
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==40.0.1
decorator==5.1.1
distlib==0.3.6
exceptiongroup==1.0.4
execnet==1.9.0
filelock==3.8.0
firebase-admin==6.1.0
frozenlist==1.3.3
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-appengine-logging==1.2.0
google-cloud-audit-log==0.2.4
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-logging==3.3.1
google-cloud-secret-manager==2.16.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpc-google-iam-v1==0.12.4
grpcio==1.53.0
grpcio-status==1.53.0
gunicorn==20.1.0
httplib2==0.21.0
identify==2.5.22
idna==3.4
iniconfig==1.1.1
Jinja2==3.1.2
libcst==0.4.9
MarkupSafe==2.1.2
model-bakery==1.9.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==0.4.3
nodeenv==1.7.0
numpy==1.24.3
oauth2client==4.1.3
openai==0.27.7
packaging==22.0
platformdirs==3.2.0
pluggy==1.0.0
pre-commit==2.20.0
proto-plus==1.22.2
protobuf==4.22.3
psycopg==3.1.6
psycopg-pool==3.1.7
psycopg2==2.9.3
psycopg2-binary==2.9.6
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytest-env==0.8.1
pytest-forked==1.4.0
pytest-split==0.8.0
pytest-xdist==2.5.0
python-dotenv==1.0.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tabulate==0.9.0
toml==0.10.2
tomli==2.0.0
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.21.0
wait-for-it==2.2.1
whitenoise==6.2.0
wrapt==1.15.0
yarl==1.9.2
